Add initial CRUD API endpoints (#15)
* Add get_items API and basic display for frontend

`get_items()` currently returns an unfiltered list of all items in db.
Will need to figure out if we want to pass in some sort of filter
function or delegate filtering to a seperate function.

Added some placeholder display logic to our main view - nothing
permanent, just for testing.

* add filtering to get_items

* allow multiple filters to be passed to get_items

* refactor commands, schema, and filtering to modules

* add singular get_item command

* add seed_db

* adjust async funs

* adjust naming

* Add create_item API

Can take in an optional rating and category, defaulting to 0 if rating
isn't present. This should simplify the amount of functions we need at
the callsite whilst still giving us flexibility.

* change schema

* add defaults to schema, add delete_db script

* get and delete categories


* update error return formatting

* Add get_items API and basic display for frontend

`get_items()` currently returns an unfiltered list of all items in db.
Will need to figure out if we want to pass in some sort of filter
function or delegate filtering to a seperate function.

Added some placeholder display logic to our main view - nothing
permanent, just for testing.

* add filtering to get_items

* allow multiple filters to be passed to get_items

* refactor commands, schema, and filtering to modules

* add singular get_item command

* add seed_db

* adjust async funs

* adjust naming

* bugfix

* Add create_item API

Can take in an optional rating and category, defaulting to 0 if rating
isn't present. This should simplify the amount of functions we need at
the callsite whilst still giving us flexibility.

* change schema

* add defaults to schema, add delete_db script

* get and delete categories

* update error return formatting

* add get_ratings

* add DELETE CASCADE to schema

* remove unneccesary DELETE query

* add delete_rating

* change schema to use named pk's and chrono datetime

* add update item and switch id naming conventions

* initial testing

* move db logic to db folder

* split up create item into seperate db queries

* update category

* update rating and delete item

---------

Co-authored-by: Xinrui Chen <xyn.hello@gmail.com>